+++ src/scanner/ScannerResult.js
@@ -37,6 +37,9 @@ export function createScannerResult(file) {
     snapshot: '',
     value: '',
     status: '',
+    takeId: '',
+    /** @type {{ projectId: string, shotHash: string }|null} */
+    takeInfo: null,
   };
 }
 
@@ -61,32 +64,59 @@ export function readyScanner(out, index) {
 
 /**
  * @param {ScannerResult} out
+ * @param {import('@/stores/document/DocumentStore').TakeId} takeId
+ */
+export function attachTakeIdScanner(out, takeId) {
+  out.takeId = takeId;
+  return out;
+}
+
+/**
+ * @param {ScannerResult} out
+ * @param {{ projectId: string, shotHash: string }} takeInfo
+ */
+export function attachTakeInfoScanner(out, takeInfo) {
+  out.takeInfo = takeInfo;
+  return out;
+}
+
+/**
+ * @param {ScannerResult} out
  */
 export function skipScanner(out) {
-  return failScanner(out, '[SKIP]');
+  for (let key of Object.keys(out)) {
+    // @ts-expect-error Skip should just make this look empty.
+    delete out[key];
+  }
+  out.status = '[SKIP]';
+  return out;
 }
 
 /**
  * @param {ScannerResult} out
  * @param {string} status
+ * @param {boolean} [force]
  */
-export function failScanner(out, status) {
+export function failScanner(out, status, force = true) {
   out.status = status;
-  out.file = null;
+  if (force) {
+    out.file = null;
+  }
   return out;
 }
 
 /**
  * @param {ScannerResult} out
  * @param {any} error
+ * @param {boolean} [forceFailure]
  */
-export function errorScanner(out, error) {
+export function errorScanner(out, error, forceFailure = true) {
   if (error instanceof Error) {
-    return failScanner(out, `[ERROR: ${error.message}]`);
+    return failScanner(out, `[ERROR: ${error.message}]`, forceFailure);
   } else if (typeof error === 'string') {
-    return failScanner(out, `[ERROR: ${error}]`);
+    return failScanner(out, `[ERROR: ${error}]`, forceFailure);
   } else {
-    return failScanner(out, `[ERROR: not an error? - ${error}]`);
+    return failScanner(out, `[ERROR: not an error? - ${error}]`, forceFailure);
   }
 }