+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - shot details focus experience
 - - nav takes on shotlist
 - - nav shots on edit mode
+- Be able to move shots across scenes
+- Be able to record QR code in recorder
+
+- Add import all or verified footage for scanner
+- Add take id to qr code key
+- Fix qr code aspect ratio
 
 # 0.1.14